метод _renderTasksPortion
