+++ markup/css/style.css
@@ -809,6 +809,6 @@ body {
   justify-content: space-between;
 }
 
-.hidden {
+.hide {
   display: none;
 }
