+++ src/components/data.js
@@ -13,7 +13,7 @@ const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
 const WEEK_DAYS = [`Mo`, `Tu`, `We`, `Th`, `Fr`, `Sa`, `Su`];
 
 export let TASKS = [];
-const TASKS_COUNT = 21;
+const TASKS_COUNT = 10;
 
 const MSEC_IN_A_WEEK = 7 * 24 * 60 * 60 * 1000;
 const weekBehindNow = new Date() - MSEC_IN_A_WEEK;
@@ -51,11 +51,12 @@ const getTask = () => (
   }
 );
 
-const getTasksData = () => {
-  for (let i = 0; i < TASKS_COUNT; i++) {
-    TASKS.push(getTask());
+const getTasksData = (tasksArray, tasksCount) => {
+  for (let i = 0; i < tasksCount; i++) {
+    tasksArray.push(getTask());
+    tasksArray[i].id = i;
   }
 };
 
-getTasksData();
+getTasksData(TASKS, TASKS_COUNT);
